Renamed table_generator.h to .hpp
