+++ src/lib/cost_calibration/table_generator.cpp
@@ -2,7 +2,7 @@
 // Created by Lukas Böhme on 17.12.19.
 //
 
-#include "table_generator.h"
+#include "table_generator.hpp"
 #include "storage/table.hpp"
 #include "constant_mappings.hpp"
